added compass as bookmark
Now the bookmark handling is more intuitive.
The first valid item from the right is used and
put in the first free slot from the left.
No more prioritizing of books over wands.

Also will not halt when encountering a book
not containing valid data. The next valid item
is searched for.